Use population-based weighted sampling for Asia builds
This replaces the Asia/China/India split with population-based weighted
sampling (possible in Augur version 25.3.0).

This requires changing the geographical grouping resolution from
division to country, but I assume it was only grouped by division in an
attempt to have varying group sizes per country, and that
population-based weighting is an acceptable replacement.