Don't include Wuhan root in Nextstrain profile
Drop forced inclusion of Wuhan/1 root in the Nextstrain GISAID profile and swap rooting to use "best", ie temporally optimal rooting. This allows the root to be the common ancestor of the subsampled sequences. This makes it so that with the changes to time-based subsampling in the previous commit, the "6m" analysis includes samples from the previous 30m and the TMRCA is in ~2021.

This set up should be significantly more future proof than needing to continually make new clade-specific (eg /21L/) roots as selective sweeps occur.