Add linkchecker configuration [#1133]
Note: this contains 2 URLs in the `linkcheck_ignore` section that need
to be fixed; each has an associated FIXME comment linking to a Github
issue.

I'm ignoring those temporarily so that after this PR is merged, `make
linkcheck` will exit without error in this repo, which will unblock
activating linkchecking in _all_ the CI doc builds.