Remove unused mutations summary rules & script
This commit removes the rules `mutation_summary` and
`build_mutation_summary`, both of which called the removed script
`scripts/mutation_summary.py`.

I was looking into updating the script to work with Nextclade v3 outputs
but it seems like the rules using the script are not actively used in
the Snakemake workflow!

In Feb 2021, the script and rule `mutation_summary` were added and the
only use of the output `results/mutation_summary_{origin}.tsv.xz`
was to be uploaded it as an intermediate file of the builds.¹
In Jan 2022, the upload of the intermediate file was removed² and I
cannot find any other uses of the file so the rule `mutation_summary`
seems unused.

In March 2021, the rule `build_mutation_summary` was added to support
the rule `add_mutation_counts`.³ However, `add_mutation_counts` was
later updated to use `augur distance`⁴ which does not require the
mutation summary output. I cannot find any other uses of the output file
`results/{build_name}/mutation_summary.tsv`, so the rule
`build_mutation_summary` seems unused.

¹ <https://github.com/nextstrain/ncov/pull/562>
² <https://github.com/nextstrain/ncov/pull/814>
³ <https://github.com/nextstrain/ncov/commit/3b23e8f66766a543c98827029f588c9bed2d5b36>
⁴ <https://github.com/nextstrain/ncov/commit/d24d53122b3cb6a76309a5adea95819841c8eb89>