Migrate to Nextclade v3
Major change includes the removal of the insertions.csv output from the
`build_align` rule. This output was only used by the
`build_mutation_summary` rule, which was removed in a previous commit.

Note: Using `nextclade` instead of `nextclade3` to allow the workflow to
continue to work with Snakemake's `--use-conda` interface.