Remove unused outputs for `align`
The only downstream use of the insertions and translations outputs of
`align` rule was the `mutation_summary` rule which was removed in the
previous commit.

Even though this is no longer doing translations, we are still keeping
the `annotation` config param and the `defaults/annotation.gff` file
for codon-aware nucleotide alignment.
<https://github.com/nextstrain/ncov/pull/1160#issuecomment-2433703300>