Remove commit history link
The commit history for the entire repo is not very meaningful to
represent the text that is linked. I considered scoping to the
`defaults/` folder with
<https://github.com/nextstrain/ncov/commits/master/defaults>, but that
also shows many irrelevant commits. I think it's easiest to just drop
this link.

Original motivation is CI failing on the URL rate limiting during docs
linkcheck.