Skip sequence checks
This avoids an extra pass through the sequence file as of Augur 31.2.1.

The sequence checks do 2 things: (1) check for duplicates and (2) check
for presence of ids from metadata so that any ids missing from sequences
are dropped.

Here's why it's safe to bypass these checks:

(1) is not useful since sanitize_sequences.py already drops duplicates.

(2) is not useful since --exclude-all already excludes any ids that
would have been excluded by the check for presence in sequences.