Skip sequence checks in Nextstrain configs
This avoids an extra pass through the sequence file that was introduced
in Augur 31.2.1.

The sequence checks do 2 things: (1) check for duplicates and (2) check
for presence of ids from metadata so that any ids missing from sequences
are dropped.

Here's why it's safe to bypass these checks:

(1) is not useful if the inputs are already deduplicated prior to
running this rule. A new config key has been added to mark inputs as
deduplicated.

(2) is not useful since --exclude-all already excludes any ids that
would have been excluded by the check for presence in sequences.