[snakemake] update memory calculations
As we now require snakemake v8 we can leverage the more convenient
`input.size_mb` functionality, which is functionally equivalent to
`input.size / 1024 / 1024` [1]

The `rule colors` wants the size of an individual input (not the sum of
all inputs) and the `_IOFile.size` function is now a coroutine. The more
widely used (by us) `InputFiles.size` function uses `async_run` to allow
coroutine calls so I think it's fine for us to do so explicitly to get
the size of an individual file. Other code paths end up at the same
point, such as `input.size_files_mb[list(input.keys()).index('metadata')]`

Closes #1114

[1] <https://github.com/snakemake/snakemake/blob/7f2606e453ba6bd83222ec3679178ae96bd142d2/src/snakemake/io/__init__.py#L1894-L1923>
[2] <https://github.com/snakemake/snakemake/blob/7f2606e453ba6bd83222ec3679178ae96bd142d2/src/snakemake/io/__init__.py#L622-L633>