[snakemake] remote storage plugins for path_or_url
The previous version of `path_or_url` was not compatible with Snakemake
v8, which implemented a storage plugin architecture [1]. This is a
complete rewrite of the approach and was originally written for zika
[2]. See [2] for more commentary as well as the docs added to
`remote_files.smk`

Google Storage URIs are not supported due to authn complexity, as
discussed in [3] and HTTP support is dropped as requested in [4].

The default of `keep_local` is now True. Since all (n=3) existing usages
of `path_or_url` specify `keep_local=True` this is consistent with
previous behaviour.

[1] <https://snakemake.readthedocs.io/en/stable/snakefiles/storage.html>
[2] <https://github.com/nextstrain/zika/pull/89>
[3] <https://github.com/nextstrain/zika/pull/89#discussion_r2190671313>
[4] <https://github.com/nextstrain/zika/pull/89#discussion_r2191122400>